Fixed imports to be local, not relative, and fixed circular imports
- moved handlers to main due to circular imports
- adjusted imports to be local, not relative
