+++ __init__.py
@@ -1,10 +1,10 @@
 # fractional_indexing/__init__.py
 
-from .fractional_indexing import (
+from fractional_indexing import (
     generate_key_between,
     generate_n_keys_between
 )
-from .exceptions import OrderKeyError
+from exceptions import OrderKeyError
 
 __all__ = [
     'generate_key_between',
