+++ tests.py
@@ -7,7 +7,7 @@ from fractional_indexing import (
     generate_key_between,
     generate_n_keys_between,
 )
-from .utils import validate_order_key
+from utils import validate_order_key
 
 BASE_95_DIGITS = ' !"#$%&\'()*+,-./0123456789:;<=>?@ABCDEFGHIJKLMNOPQRSTUVWXYZ[\\]^_`abcdefghijklmnopqrstuvwxyz{|}~'
 
@@ -45,13 +45,14 @@ def test_generate_key_between(a: Optional[str], b: Optional[str], expected) -> N
     if isinstance(expected, OrderKeyError):
         with pytest.raises(OrderKeyError) as e:
             generate_key_between(a, b)
-        assert e.value.args[0] == expected.args[0]
+        assert e.value.args[0].lower() == expected.args[0].lower()
     else:
         act = generate_key_between(a, b)
         print(f'exp: {expected}')
         print(f'act: {act}')
         print(act == expected)
         assert act == expected
+
 
 
 @pytest.mark.parametrize(['a', 'b', 'n', 'expected'], [
@@ -94,7 +95,7 @@ def test_base95_digits(a: Optional[str], b: Optional[str], expected: str) -> Non
     if isinstance(expected, OrderKeyError):
         with pytest.raises(OrderKeyError) as e:
             generate_key_between(**kwargs)
-        assert e.value.args[0] == expected.args[0]
+        assert e.value.args[0].lower() == expected.args[0].lower()
     else:
         act = generate_key_between(**kwargs)
         print()
@@ -145,13 +146,11 @@ def test_readme_examples_multiple_keys():
 
 
 def test_readme_examples_validate_order_key():
-    from fractional_indexing import validate_order_key, OrderKeyError
-
     validate_order_key('a0')
 
     with pytest.raises(OrderKeyError) as e:
         validate_order_key('foo')
-    assert str(e.value) == 'invalid order key: foo'
+    assert str(e.value).lower() == 'invalid order key: foo'.lower()
 
 
 def test_readme_examples_custom_base():
